post creats new account and delete removes account

+++ api/accounts/accounts-router.js
@@ -1,4 +1,4 @@
-const { checkAccountId } = require('./accounts-middleware')
+const { checkAccountId, checkAccountPaylod } = require('./accounts-middleware')
 const Accounts = require('./accounts-model')
 //{ getAll }
 const router = require('express').Router()
@@ -7,7 +7,7 @@ router.get('/', async (req, res, next) => {
   try {
     const allAccounts = await Accounts.getAll(req.body)
     return res.status(200).json(allAccounts)
-    
+
   } catch (err) {
     res.status(500).json({ message: 'Somethings wrong' })
     next(err)
@@ -17,34 +17,88 @@ router.get('/', async (req, res, next) => {
 // res.status(404).json({  });
 router.get('/:id', async (req, res, next) => {
   try {
-    const  [ fetchedAccount ] = await Accounts.getById(req.params.id)
-    if (!fetchedAccount) 
-    return res.status(404).json({message: 'account not found'})
-return res.json(fetchedAccount)
+    const [fetchedAccount] = await Accounts.getById(req.params.id)
+    if (!fetchedAccount)
+      return res.status(404).json({ message: 'account not found' })
+    return res.json(fetchedAccount)
   } catch (error) {
     res.status(500).json({ message: 'Somethings wrong' })
-    
+
   }
-next()
+  next()
 })
 
-router.post('/', (req, res, next) => {
-  // DO YOUR MAGIC
+router.post('/', async (req, res, next) => {
+
+  // try {
+  //   res.json('post account')
+  // } catch (error) {
+  //   next(error)
+  // }
+
+try {
+  const data = await Accounts.create(req.body)
+  res.json(data)
+} catch (error) {
+  next(error)
+}
+// try {
+//   const { name, budget } = req.body;
+// const newAccount = Accounts.insert({name , budget})
+// if(!name || !budget){
+//   res.status(400).json({message: "missing some stuff"})
+// } else{
+//   res.status(201).json(newAccount)
+// }
+
+// } catch (error) {
+
+// }
+
 })
 
 router.put('/:id', (req, res, next) => {
-  // DO YOUR MAGIC
+  const  changes  = req.body;
+  Accounts.updateById(req.params.id, changes)
+  .then(updatedAccount => {
+    if(!updatedAccount) {
+      res.status(404).json({message: 'Doesnt exist'})
+    } else if (!updatedAccount.name || !updatedAccount.budget) {
+      res.status(400).json({message: 'Need some info bucko'})
+    } else{
+      res.status(200).json(updatedAccount)
+    }
+  })
+  .catch(() => {
+    next()
+    // res.status(500).json({message: 'We Need a new server'})
+  })
+
+
+
+
+  // try {
+  //   res.json('update account')
+  // } catch (error) {
+  //   next(error)
+  // }
 });
 
-router.delete('/:id', checkAccountId, async (req, res, next) => {
+router.delete('/:id',  async (req, res, next) => {
   try {
-    const killSwitch = await Accounts.deleteById(req.params.id)
-    res.json(killSwitch)
+    const killSwitch = await Accounts.getById(req.params.id)
+    if (!killSwitch) {
+      res.status(404).json({ message: 'whoa! that ID doesnt exist' })
+    } else {
+      await Accounts.deleteById(req.params.id)
+      res.status(404).json({ message: 'account not found' })
+    }
+    // res.json(killSwitch)
   } catch (error) {
-    next(error)
-    // res.status(500)
+    
+    res.status(500)
   }
-
+next()
 
 })
